Update example endpoints to use new structure
