+++ api/package.json
@@ -39,7 +39,7 @@
     "@aws-sdk/client-ecr": "^3.229.0",
     "@aws-sdk/client-s3": "^3.180.0",
     "@aws-sdk/lib-dynamodb": "^3.234.0",
-    "@connected-web/openapi-rest-api": "^0.1.4",
+    "@connected-web/openapi-rest-api": "^0.3.0",
     "ajv": "^8.12.0",
     "ajv-errors": "^3.0.0",
     "ajv-formats": "^2.1.1",
